+++ scripts/t.js
@@ -1,3 +1,7 @@
-export function t(key) {
-  return browser.i18n.getMessage(key) || key;
+export function t(key1, key2) {
+  if (key2) {
+    return browser.i18n.getMessage(key1, key2) || key1;
+  }
+  
+  return browser.i18n.getMessage(key1) || key1;
 }