+++ update/update.js
@@ -4,8 +4,9 @@ document.addEventListener('DOMContentLoaded', () => {
   document.getElementById('version').textContent = version;
   
   const features = [
-    'Implemented a new link in the popup UI that directs users to the web extension \'s settings page.',
-    'Enhanced user experience by providing faster access to extension settings.'
+    'We’ve made several improvements to stability and accessibility.',
+    'The issue with closing tabs by rule has been fixed, and translations with placeholders now display correctly.',
+    'We also improved the handling of localized messages in the settings page and added accessibility attributes to form inputs and table headers.'
   ];
   
   const ul = document.getElementById('features');